requirements.txt
requests
scrapy
pandas
xlwt
openpyxl
xlrd
wordcloud
jieba
numpy
pillow
pyecharts
snapshot_selenium
echarts-countries-pypkg
echarts-china-provinces-pypkg
echarts-china-cities-pypkg
echarts-china-counties-pypkg
wxpy
scrapy